### 2021-07-12 #3486 @kthyng
@xp-song I think your paper needs to be named `paper.md` to be found. Mote info [here](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain).
