### 2021-07-12 #3486 @kthyng
Since this software is on the small side (<1000 lines of code), I am pinging the editorial board to make sure it is in scope in terms of [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort). This process will take 1-2 weeks.
