### 2021-07-14 2b630ad
- The use of `terra::extract()` may output an additional third `area` column that is unused (all `NA`s). Only keep first 2 columns of output.
