+++ NEWS.md
@@ -1,4 +1,12 @@
 
+# home2park (development version)
+
+
+## Bug fixes 
+* Helper function `raster_class_area()`. The use of `terra::extract()` may output an additional third `area` column that is unused. Only keep first 2 columns of output.
+
+<br>
+
 # home2park 0.1.0
 
 * Finish development of main functions
