### 2021-07-18 06f7047
edit name of JOSS paper to paper.md
