### 2021-07-18 #3486 @xp-song
> @xp-song I think your paper needs to be named `paper.md` to be found. Mote info [here](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain).

Hi @kthyng, thank you for looking into our work. I've fixed this in commit [06f7047](https://github.com/ecological-cities/home2park/commit/06f70477262169855daa371c552f10057e5559a7)
