### 2021-07-18 #3486 @xp-song
> @xp-song if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).

Here are several reviewer suggestions based on the provided list:

- Jeffrey Hollister ([jhollist](https://github.com/jhollist)) - work on spatial ecology and environmental sciences, and experience in GIS and spatial statistics in R.
- Bryce Mecum ([amoeba](https://github.com/amoeba)) - experience in spatial data manipulation in R.
- Daniel Nüst ([nuest](https://github.com/nuest)) - experience in GIS and spatiotemporal data in R.
- Virgilio Gómez-Rubio ([becarioprecario](https://github.com/becarioprecario)) - work on GIS and spatial statistics in R.
