### 2021-07-18 #3486 @xp-song
@danielskatz The package can be used anywhere (user provided data), and functions are included to download/process OSM data. So out of the box, it may be used anywhere where OSM is available. The example provided is for the city of Singapore
