### 2021-07-22 ff2b933
fix bug in rasterise_buildings(), which imported names for intermediate (processing) land use rasters wrongly, and also renamed them unsuccessfully because of glue::glue() function working differently within package environment
