### 2021-08-04 d964c8d
add LazyDataCompression type in DESCRIPTION
