### 2021-08-04 43312e7
Add removal of invalid building polygons in get_buildings_osm(), which remained in spite of st_make_valid(). Related to https://github.com/r-spatial/sf/issues/1649
