### 2021-08-04 #3486 @xp-song
@danielskatz The research areas that [home2park](https://ecological-cities.github.io/home2park/) intersect with include topics such as GIS/geospatial analysis (e.g. spatial equity/statistics), human ecology (e.g. human–nature interactions, recreation/leisure), as well as city/urban planning (e.g. landscape ecology/patterns). Some suggested editors are listed below, hope this helps:

- Kristina Riemer ([KristinaRiemer](https://github.com/KristinaRiemer)) - background in environmental science, and ecology with extensive experience in R (spatial data, ecosystem modelling)
- Marcos Vital ([@marcosvital](https://github.com/marcosvital)) - extensive work in R and ecology
- Hugo Ledoux ([hugoledoux)](https://github.com/hugoledoux)) - extensive work on GIS, geoinformation and their application
