### 2021-08-04 936f7c0
add customisable arguments for sf::st_write() within get_buildings_osm().
- driver
- delete_dsn
