+++ NEWS.md
@@ -7,7 +7,7 @@
 
 * Fix error in `rasterise_buildings()`. Getting the files for land use rasters resulted in an error (e.g. multiple files). Syntax for re-naming for the output using the `glue::glue()` function also did not work within package environment.
 
-* Add removal of invalid building polygons in `get_buildings_osm()`, which remained in spite of `st_make_valid()`. Related to https://github.com/r-spatial/sf/issues/1649
+* In `get_buildings_osm()`, add removal of invalid building polygons which remained in spite of `st_make_valid()`. Related to https://github.com/r-spatial/sf/issues/1649. Also, add arguments for `sf::st_write()` (e.g. `driver`, `delete_dsn`).
 
 <br>
 
