### 2021-08-04 a70f9c1
add customisable arguments for sf::st_write() within get_buildings_osm().
- append
