+++ R/get_buildings_osm.R
@@ -19,11 +19,12 @@
 #'for the specific dates available. Defaults to `NULL` (download the latest available data).
 #'@param dir_raw character. Directory to download the raw unprocessed OSM data. Passed to
 #'argument `download_directory` in `osmextract::oe_read()`.
+#'@param ... Other arguments passed to `osmextract::oe_read()`.
 #'@param filename character (optional). File path to export output data.
 #'@param driver character (optional). Name of driver used to export output data, passed to `sf::st_write()`.
 #'Defaults to "GeoJSON".
 #'@param delete_dsn logical (optional). Passed to `sf::st_write()`.
-#'@param ... Other arguments passed to `osmextract::oe_read()`.
+#'@param append defaults to `NA`, which raises an error if a layer exists. Passed to `sf::st_write()`.
 #'
 #'@return The processed building polygons (`sf` object).
 #'
@@ -57,7 +58,7 @@
 #'
 #'@export
 get_buildings_osm <- function(place, date = NULL, dir_raw = osmextract::oe_download_directory(), filename = NULL, driver = "GeoJSON", delete_dsn = TRUE,
-    ...) {
+                              append = NA, ...) {
 
     # Error checking ------------------
 
@@ -131,7 +132,7 @@ get_buildings_osm <- function(place, date = NULL, dir_raw = osmextract::oe_downl
 
     # export ----
     if (!is.null(filename)) {
-        sf::st_write(results, filename, driver = driver, delete_dsn = delete_dsn)
+        sf::st_write(results, filename, driver = driver, delete_dsn = delete_dsn, append = append)
     }
 
     rm(link, osmkeys, q, bb)
