### 2021-08-04 4629045
remove empty polygons in get_buildings_osm() after st_make_valid(), as they are still present
