+++ R/get_buildings_osm.R
@@ -106,9 +106,6 @@ get_buildings_osm <- function(place, date = NULL, dir_raw = osmextract::oe_downl
 
     # clean up ---- remove empty geoms, transform back to same crs as 'place', cast to
     # individual polygons
-
-
-
     suppressWarnings(results <- results %>%
         dplyr::filter(!sf::st_is_empty(.)) %>%
         sf::st_transform(sf::st_crs(place)) %>%
@@ -116,7 +113,7 @@ get_buildings_osm <- function(place, date = NULL, dir_raw = osmextract::oe_downl
 
     # rm invalid building polygons
     results <- results[sf::st_is_valid(results),]
-    results <- results[!st_is_empty(results),]
+    results <- results[!sf::st_is_empty(results),]
 
     results <- results %>%
         sf::st_cast("MULTIPOLYGON", warn = FALSE) %>%
