### 2021-08-05 05b6041
bug fix for rasterise_buildings
- 'year' argument default value changed to NULL
