### 2021-08-05 732e9fb
Add section to rasterise buildings without information on the 'year'. Assumes that only one population/landuse rasters were generated in pre-processing step too; if there are multiple, uses the first object for analysis. Also, add  `st_crs()` checks (e.g. similar crs between `sf_buildings` and `sf_pop` if supplied; similar crs between `sf_buildings` and `rastertemplate`)
