+++ R/rasterise_buildings.R
@@ -102,9 +102,10 @@ rasterise_buildings <- function(sf_buildings, proxy_pop_density = NULL, year = N
     coll <- checkmate::makeAssertCollection()
 
     # file formats
-    checkmate::assertTRUE(!st_is_longlat(sf_buildings) & !is.null(st_crs(sf_buildings)), add = coll)  # must be projected crs
+    checkmate::assertTRUE(!is.null(st_crs(sf_buildings)), add = coll)  # must have crs
     # checkmate::assertTRUE(all(st_is_valid(sf_buildings)), add = coll) # all features must be
     # valid (too time consuming)
+
 
     # colnames
     checkmate::assert_subset(proxy_pop_density, choices = colnames(sf_buildings), empty.ok = FALSE,
@@ -133,6 +134,8 @@ rasterise_buildings <- function(sf_buildings, proxy_pop_density = NULL, year = N
     # IF POP DATA IS PROVIDED
     if (!is.null(sf_pop)) {
 
+        checkmate::assertTRUE(st_crs(sf_buildings) == st_crs(sf_pop))  # must have same crs
+
         # match pop data to most recent building data
         sf_pop <- matchyear(data = sf_pop, data_tomatch = sf_buildings, year = year, match = match_buildings_pop)
 
@@ -151,8 +154,10 @@ rasterise_buildings <- function(sf_buildings, proxy_pop_density = NULL, year = N
     # GET RASTER TEMPLATE IF USER SUPPLIED
     if (!is.null(dir_rastertemplate)) {
         raster_template <- rast(dir_rastertemplate)
+        checkmate::assertTRUE(st_crs(sf_buildings) == st_crs(raster_template))  # must have same crs
     } else {
         raster_template <- rast(glue::glue("{dir_processing}/popdensity_raster-template.tif"))
+        checkmate::assertTRUE(st_crs(sf_buildings) == st_crs(raster_template))  # must have same crs
     }
 
 
@@ -160,123 +165,193 @@ rasterise_buildings <- function(sf_buildings, proxy_pop_density = NULL, year = N
     results <- list()  # create empty list for output
 
     # If there is a year column in sf_buildings
-    if (!is.null(unique(sf_buildings[[year]])))
-        {
+    if (!is.null(year)) {
 
-            years_buildings <- unique(sf_buildings[[year]])
+        years_buildings <- unique(sf_buildings[[year]])
 
-            # each buildings year
-            for (i in seq_along(years_buildings)) {
+        # each buildings year
+        for (i in seq_along(years_buildings)) {
 
-                sf_buildings_subset <- sf_buildings %>%
-                  dplyr::filter(.data[[year]] == years_buildings[i])
+            sf_buildings_subset <- sf_buildings %>%
+              dplyr::filter(.data[[year]] == years_buildings[i])
 
-                # rasterize & export to temp directory
-                buildings_raster <- terra::rasterize(terra::vect(sf_buildings_subset), raster_template,
-                  field = proxy_pop_density)
+            # rasterize & export to temp directory
+            buildings_raster <- terra::rasterize(terra::vect(sf_buildings_subset), raster_template,
+              field = proxy_pop_density)
 
-                buildings_raster[buildings_raster == 0] <- NA  # convert 0s to NAs
-
-
-                # IF POP DATA IS PROVIDED, remove areas not covered by pop census blocks (for matching
-                # census yr)
-                if (!is.null(sf_pop)) {
-
-                  years_popmatch <- unique(sf_pop$year[sf_pop$year_match == years_buildings[i]])
-                  years_popmatch <- stats::na.omit(years_popmatch)
-
-                  results[[i]] <- list()
-
-                  for (j in seq_along(years_popmatch)) {
-
[…]
