### 2021-08-05 1e52777
in rasterise_buildings(), allow raster template (`terra::rast()` object) to be input directly as argument, or alternatively supply file path to the raster file.
