+++ R/rasterise_buildings.R
@@ -35,8 +35,8 @@
 #'@param match_buildings_landuse character. Type of matching between `sf_buildings` and `sf_landuse` (if provided), passed on to `match` argument in function `matchyear()`.
 #'Either `'exact'`, `'closest'`, `'recent'` or `'soonest'`. Defaults to `'closest'`.
 #'This argument is used only if `sf_pop` is not provided.
-#'@param dir_rastertemplate character. Filepath to the raster used to define the pixel resolution, extent, nrow, ncol of
-#'the output raster; object is passed to the '`y`' argument in `terra::rasterize()`.
+#'@param raster_template Either a `terra::rast` object, or a filepath to the raster used to define the pixel resolution, extent, nrow, ncol of
+#'the output raster. Object is passed to the '`y`' argument in `terra::rasterize()`.
 #'Defaults to the template raster generated by the function `rasterise_pop()` within `dir_processing` (see next argument).
 #'@param dir_processing character. Directory to get intermediate files generated in previous steps (e.g. rasterised population/land use data).
 #'Defaults to `tempdir()`.
@@ -94,7 +94,7 @@
 #'@export
 rasterise_buildings <- function(sf_buildings, proxy_pop_density = NULL, year = NULL, sf_pop = NULL,
     sf_landuse = NULL, match_buildings_pop = "closest", match_buildings_landuse = "closest",
-    dir_rastertemplate = NULL, dir_processing = tempdir(), dir_export = NULL, overwrite = TRUE,
+    raster_template = NULL, dir_processing = tempdir(), dir_export = NULL, overwrite = TRUE,
     wopt = list(gdal = c("COMPRESS=LZW")), ...) {
 
     # Error checking ------------------
@@ -119,8 +119,6 @@ rasterise_buildings <- function(sf_buildings, proxy_pop_density = NULL, year = N
         add = coll)  # as vector
 
     # file paths
-    checkmate::assert_character(dir_rastertemplate, min.len = 1, any.missing = FALSE, all.missing = FALSE,
-        null.ok = TRUE, add = coll)
     checkmate::assert_character(dir_processing, min.len = 1, any.missing = FALSE, all.missing = FALSE,
         null.ok = TRUE, add = coll)
     checkmate::assert_character(dir_export, min.len = 1, any.missing = FALSE, all.missing = FALSE,
@@ -151,14 +149,14 @@ rasterise_buildings <- function(sf_buildings, proxy_pop_density = NULL, year = N
     }
 
 
-    # GET RASTER TEMPLATE IF USER SUPPLIED
-    if (!is.null(dir_rastertemplate)) {
-        raster_template <- rast(dir_rastertemplate)
-        checkmate::assertTRUE(st_crs(sf_buildings) == st_crs(raster_template))  # must have same crs
-    } else {
+    # GET RASTER TEMPLATE (depending on object supplied)
+    if(is.null(raster_template)) {
         raster_template <- rast(glue::glue("{dir_processing}/popdensity_raster-template.tif"))
-        checkmate::assertTRUE(st_crs(sf_buildings) == st_crs(raster_template))  # must have same crs
+    } else if(!is.null(raster_template) & is.character(raster_template)) { # user supplied file path
+        raster_template <- rast(raster_template)
     }
+    checkmate::assertTRUE(st_crs(sf_buildings) == st_crs(raster_template))  # must have same crs
+
 
 
     # RASTERIZE
