+++ man/rasterise_buildings.Rd
@@ -12,7 +12,7 @@ rasterise_buildings(
   sf_landuse = NULL,
   match_buildings_pop = "closest",
   match_buildings_landuse = "closest",
-  dir_rastertemplate = NULL,
+  raster_template = NULL,
   dir_processing = tempdir(),
   dir_export = NULL,
   overwrite = TRUE,
@@ -44,8 +44,8 @@ Either \code{'exact'}, \code{'closest'}, \code{'recent'} or \code{'soonest'}. De
 Either \code{'exact'}, \code{'closest'}, \code{'recent'} or \code{'soonest'}. Defaults to \code{'closest'}.
 This argument is used only if \code{sf_pop} is not provided.}
 
-\item{dir_rastertemplate}{character. Filepath to the raster used to define the pixel resolution, extent, nrow, ncol of
-the output raster; object is passed to the '\code{y}' argument in \code{terra::rasterize()}.
+\item{raster_template}{Either a \code{terra::rast} object, or a filepath to the raster used to define the pixel resolution, extent, nrow, ncol of
+the output raster. Object is passed to the '\code{y}' argument in \code{terra::rasterize()}.
 Defaults to the template raster generated by the function \code{rasterise_pop()} within \code{dir_processing} (see next argument).}
 
 \item{dir_processing}{character. Directory to get intermediate files generated in previous steps (e.g. rasterised population/land use data).
