+++ NEWS.md
@@ -9,7 +9,11 @@
 
 * In `get_buildings_osm()`, add removal of invalid building polygons which remained in spite of `st_make_valid()`. Related to https://github.com/r-spatial/sf/issues/1649. Also, add arguments for `sf::st_write()` (e.g. `driver`, `delete_dsn`, `append`).
 
-* In `rasterise_buildings_osm()`, change default argument for 'year' to `NULL`, as multiple years may not necessarily be analysed. Add section to rasterise buildings without information on the 'year'. Assumes that only one population/landuse rasters were generated in pre-processing step too; if there are multiple, uses the first object for analysis. Also, add  `st_crs()` checks (e.g. similar crs between `sf_buildings` and `sf_pop` if supplied; similar crs between `sf_buildings` and `rastertemplate`)
+* In `rasterise_buildings_osm()`
+
+    - Change default argument for 'year' to `NULL`, as multiple years may not necessarily be analysed. Add section to rasterise buildings without information on the 'year'. Assumes that only one population/landuse rasters were generated in pre-processing step too; if there are multiple, uses the first object for analysis.
+    - Allow raster template (`terra::rast()` object) to be input directly as argument, or alternatively supply file path to the raster file.
+    - Check if CRS are similar between supplied objects (e.g. similar crs between `sf_buildings` and `sf_pop` if supplied; similar crs between `sf_buildings` and `rastertemplate`).
 
 <br>
 
