### 2021-08-05 f60e412
minor edit to rasterise_buildings() explanation of argument 'raster_template'
