+++ R/rasterise_buildings.R
@@ -36,7 +36,8 @@
 #'Either `'exact'`, `'closest'`, `'recent'` or `'soonest'`. Defaults to `'closest'`.
 #'This argument is used only if `sf_pop` is not provided.
 #'@param raster_template Either a `terra::rast` object, or a filepath to the raster used to define the pixel resolution, extent, nrow, ncol of
-#'the output raster. Object is passed to the '`y`' argument in `terra::rasterize()`.
+#'the output raster. Defaults to raster template in `tempdir()` processed in previous steps.
+#'Object is passed to the '`y`' argument in `terra::rasterize()`.
 #'Defaults to the template raster generated by the function `rasterise_pop()` within `dir_processing` (see next argument).
 #'@param dir_processing character. Directory to get intermediate files generated in previous steps (e.g. rasterised population/land use data).
 #'Defaults to `tempdir()`.
