+++ man/rasterise_buildings.Rd
@@ -45,7 +45,8 @@ Either \code{'exact'}, \code{'closest'}, \code{'recent'} or \code{'soonest'}. De
 This argument is used only if \code{sf_pop} is not provided.}
 
 \item{raster_template}{Either a \code{terra::rast} object, or a filepath to the raster used to define the pixel resolution, extent, nrow, ncol of
-the output raster. Object is passed to the '\code{y}' argument in \code{terra::rasterize()}.
+the output raster. Defaults to raster template in \code{tempdir()} processed in previous steps.
+Object is passed to the '\code{y}' argument in \code{terra::rasterize()}.
 Defaults to the template raster generated by the function \code{rasterise_pop()} within \code{dir_processing} (see next argument).}
 
 \item{dir_processing}{character. Directory to get intermediate files generated in previous steps (e.g. rasterised population/land use data).
