### 2021-08-05 64a05de
remove the need for pop_dasymap() to refer to raster_template (refer to raster from argument 'land_relative_density' instead)
