+++ R/pop_dasymap.R
@@ -14,9 +14,6 @@
 #'@param land_relative_density Raster of the relative density (e.g. suitability, habitability) of land for a single time period.
 #'May be extracted from results generated by `rasterise_buildings()` or `rasterise_landuse()`.
 #'@param filename character (optional). Export output raster to disk.
-#'@param dir_rastertemplate character. Filepath to the raster used to define the pixel resolution, extent, nrow, ncol of
-#'the output raster; object is passed to the '`y`' argument in `terra::rasterize()`.
-#'Defaults to the template raster generated by the function `rasterise_pop()` within the `tempdir()`.
 #'@param overwrite logical. Argument passed to `terra::writeRaster()`. If `TRUE`, `filename` is overwritten.
 #'@param ... Other arguments passed to `terra::writeRaster()`.
 #'
@@ -88,14 +85,14 @@
 #'
 #'@export
 pop_dasymap <- function(pop_polygons, pop_perblock_count, pop_perblock_density, land_relative_density,
-    filename = NULL, dir_rastertemplate = NULL, overwrite = TRUE, ...) {
+    filename = NULL, overwrite = TRUE, ...) {
 
 
     # Error checking ------------------
 
     coll <- checkmate::makeAssertCollection()
 
-    # file formats
+    # file formats for polygons
     checkmate::assertTRUE(!st_is_longlat(pop_polygons) & !is.null(st_crs(pop_polygons)), add = coll)  # must be projected crs
     checkmate::assertTRUE(all(st_is_valid(pop_polygons)), add = coll)  # all features must be valid
 
@@ -106,21 +103,11 @@ pop_dasymap <- function(pop_polygons, pop_perblock_count, pop_perblock_density, 
     # file paths
     checkmate::assert_character(filename, min.len = 1, any.missing = FALSE, all.missing = FALSE,
         null.ok = TRUE, add = coll)
-    checkmate::assert_character(dir_rastertemplate, min.len = 1, any.missing = FALSE, all.missing = FALSE,
-        null.ok = TRUE, add = coll)
 
     checkmate::reportAssertions(coll)
 
 
     # Calculations ------------------
-
-    # GET RASTER TEMPLATE IF USER SUPPLIED
-    if (!is.null(dir_rastertemplate)) {
-        raster_template <- rast(dir_rastertemplate)
-    } else {
-        raster_template <- rast(paste0(tempdir(), "/popdensity_raster-template.tif"))
-    }
-
 
     # 1) Get relative densities per census block polygon multiply uniform pop density per
     # census block with relative density grid (suitable pixels & no. of levels for pop data to
@@ -142,7 +129,7 @@ pop_dasymap <- function(pop_polygons, pop_perblock_count, pop_perblock_density, 
 
 
     # rasterize
-    pop_perblock_rel_density_sum <- terra::rasterize(vect(pop_perblock_rel_density_sum), raster_template,
+    pop_perblock_rel_density_sum <- terra::rasterize(vect(pop_perblock_rel_density_sum), land_relative_density,
         field = "sum")
 
 
