+++ man/pop_dasymap.Rd
@@ -10,7 +10,6 @@ pop_dasymap(
   pop_perblock_density,
   land_relative_density,
   filename = NULL,
-  dir_rastertemplate = NULL,
   overwrite = TRUE,
   ...
 )
@@ -29,10 +28,6 @@ May be extracted from the \code{pop_density_rasters} list element generated by \
 May be extracted from results generated by \code{rasterise_buildings()} or \code{rasterise_landuse()}.}
 
 \item{filename}{character (optional). Export output raster to disk.}
-
-\item{dir_rastertemplate}{character. Filepath to the raster used to define the pixel resolution, extent, nrow, ncol of
-the output raster; object is passed to the '\code{y}' argument in \code{terra::rasterize()}.
-Defaults to the template raster generated by the function \code{rasterise_pop()} within the \code{tempdir()}.}
 
 \item{overwrite}{logical. Argument passed to \code{terra::writeRaster()}. If \code{TRUE}, \code{filename} is overwritten.}
 
