### 2021-08-11 #3486 @martinfleis
👋 Hi @xp-song, I'll be editing your paper.

While I try to get in touch with potential reviewers, can you expand the paper to follow the guidelines?

- The Summary section of your paper should summarise the package. In its current form, it serves as an introduction and a problem statement more than a summary. It should contain a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience.
- Add State of the field section. You should describe how this software compares to other commonly-used packages, at least within the R ecosystem (and potentially other).

See the [submitting guidelines](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) and a [review checklist](https://joss.readthedocs.io/en/latest/review_checklist.html#software-paper) for details.

Thank you!
