### 2021-08-11 fda4f1a
update JOSS paper based on https://github.com/openjournals/joss-reviews/issues/3486#issuecomment-896790230
