+++ JOSS-article/paper.bib
@@ -1,38 +1,3 @@
-@misc{Lovelace2019,
-author = {Lovelace, Robin},
-title = {osmextract},
-url = {https://cran.r-project.org/web/packages/osmextract/index.html},
-year = {2019}
-}
-@article{Tan2017,
-abstract = {Urban parks are widely-recognized to provide multiple social and ecological benefits, but an increasingly number of studies show that such benefits are often inequitably distributed across socio-economic and ethnic groups. This has led to a growing interest in assessing the spatial distribution and access to parks and other green spaces as an aspect of environmental justice. Even though such spatial studies require assessment at multiple scales, how results may be influenced by scale has not received adequate attention. This study assessed the effects of scale on park provision and spatial equity in Singapore. A range of park provision indicators were used to evaluate how they are affected by scale and to explore their correlations with wealth, income and ethnicity. Scale effects were assessed using planning units adopted for national land use planning, namely, region, planning area and subzone. Scale significantly affected park provision indicators, particularly for indicators that incorporate population. Correlations at larger scale tend to be stronger than at smaller scales. There were higher park provision and lower potential for park congestion in planning units with higher wealth and income. Inequity also appears stronger when studied at smaller spatial scales. Results also showed that park provision indicators are prone to skewed frequency distributions, especially at smaller scales. The implications of the results were discussed in relation to spatial equity assessment and possible causes of the disparities. We also highlight the need to direct park planning at smaller scale of neighbourhoods rather than at the town or regions levels.},
-author = {Tan, Puay Yok and Samsudin, Rosita},
-doi = {10.1016/j.landurbplan.2016.11.001},
-file = {:Users/xpsong/Documents/Mendeley Desktop/Tan, Samsudin/Landscape and Urban Planning/Tan, Samsudin - 2017 - Effects of spatial scale on assessment of spatial equity of urban park provision.pdf:pdf},
-issn = {01692046},
-journal = {Landscape and Urban Planning},
-keywords = {Frequency distribution,Park provision indicators,Singapore,Spatial equity,Spatial scale,Urban park},
-pages = {139--154},
-publisher = {Elsevier B.V.},
-title = {{Effects of spatial scale on assessment of spatial equity of urban park provision}},
-url = {http://dx.doi.org/10.1016/j.landurbplan.2016.11.001},
-volume = {158},
-year = {2017}
-}
-@article{Song2020a,
-author = {Song, Xiao Ping and Richards, Daniel R and He, Peijun and Tan, Puay Yok},
-doi = {10.1016/j.landurbplan.2020.103908},
-file = {:Users/xpsong/Documents/Mendeley Desktop/Song et al/Landscape and Urban Planning/Song et al. - 2020 - Does geo-located social media reflect the visit frequency of urban parks A city-wide analysis using the count and c.pdf:pdf},
-issn = {01692046},
-journal = {Landscape and Urban Planning},
-month = {nov},
-pages = {103908},
-publisher = {Elsevier},
-title = {{Does geo-located social media reflect the visit frequency of urban parks? A city-wide analysis using the count and content of photographs}},
-url = {https://linkinghub.elsevier.com/retrieve/pii/S0169204619315014},
-volume = {203},
-year = {2020}
-}
 @article{Dmowska,
[…]
