### 2021-08-12 #3609 @whedon
[REVIEW]: home2park: An R package to assess the spatial provision of urban parks
**Submitting author:** @xp-song (<a href="http://orcid.org/0000-0002-8825-195X">Xiao Ping Song</a>)
**Repository:** <a href="https://github.com/ecological-cities/home2park" target ="_blank">https://github.com/ecological-cities/home2park</a>
**Version:** v0.1.1
