### 2021-08-12 #3609 @martinfleis
👋  @xp-song @aelissa @nickbearman this is the review thread for the paper. All of our communications will happen here from now on.
