### 2021-08-28 ecological-cities/home2park#1 @nickbearman
Broken link on home page
The link to 'Get started'  (under Using home2park) to https://github.com/ecological-cities/home2park/blob/main/articles/home2park.html doesn't work - I get a 404 message

Thanks very much.

openjournals/joss-reviews#3609
