### 2021-08-28 #3609 @nickbearman
I have a query reinstallation.
On my own PC I had to install Rtools (which was fine) and then reinstall a couple of specific packages (ellipsis and backports) but I think this is just the usual R package fiddling around with that we need to do. I did get the package working on my PC.

To verify that it should do a clean install, I tried installing it on RStudio.Cloud. That, however, did not work:
`Installing package into ‘/cloud/lib/x86_64-pc-linux-gnu-library/4.1’
(as ‘lib’ is unspecified)
* installing *source* package ‘home2park’ ...
** using staged installation
** R
** data
*** moving datasets to lazyload DB
Killed
Warning message:
In i.p(...) :
  installation of package ‘/tmp/RtmpDVcFS8/filee043922ee/home2park_0.1.0.tar.gz’ had non-zero exit status`

Given this needs to be complied, does anyone know if that should be possible on RStudio.Cloud?
I have no other way of testing install on a 'clean' environment! Thanks
