### 2021-08-28 #3609 @nickbearman
Currently I can't really test the software without a link to the 'Get started' tutorial.
I have run through the README instructions, but they are quite high-level.
The paper could expand slightly on the statistic the library generates - this is not 100% clear. Possibly a list of statistics it generates might be useful? I'm not sure currently, maybe the tutorial will help. I will revisit this in a few days time and see if it is any clearer.
