### 2021-08-29 #3609 @xp-song
> Currently I can't really test the software without a link to the 'Get started' tutorial.
> I have run through the README instructions, but they are quite high-level.

Thanks for pointing out the dead link, I've amended it in commit [91e1f09](https://github.com/ecological-cities/home2park/commit/91e1f098e2671c2eb3103f96db023da6f9c3bf81). The website also contains more detailed explanations of the package: https://ecological-cities.github.io/home2park/

I'll have a look at testing the installation on other environments and see if it works, perhaps using [conda](https://stackoverflow.com/questions/45383754/simplest-way-to-test-installation-of-r-package-on-fresh-r-install)
