### 2021-09-01 #3609 @xp-song
> I have a query reinstallation.
> On my own PC I had to install Rtools (which was fine) and then reinstall a couple of specific packages (ellipsis and backports) but I think this is just the usual R package fiddling around with that we need to do. I did get the package working on my PC.
>
> To verify that it should do a clean install, I tried installing it on RStudio.Cloud. That, however, did not work:
> `Installing package into ‘/cloud/lib/x86_64-pc-linux-gnu-library/4.1’
> (as ‘lib’ is unspecified)
>
> * installing _source_ package ‘home2park’ ...
>   ** using staged installation
>   ** R
>   ** data
>   *** moving datasets to lazyload DB
>   Killed
>   Warning message:
>   In i.p(...) :
>   installation of package ‘/tmp/RtmpDVcFS8/filee043922ee/home2park_0.1.0.tar.gz’ had non-zero exit status`
>
> Given this needs to be complied, does anyone know if that should be possible on RStudio.Cloud?
> I have no other way of testing install on a 'clean' environment! Thanks

I'm not too experience with testing with 'clean' installation myself (other than the [R-CMD-Check](https://github.com/r-quantities/units/issues/223), which works fine), but based on [this post](https://stackoverflow.com/questions/45383754/simplest-way-to-test-installation-of-r-package-on-fresh-r-install), I managed to install it using a new conda environment:

**In my command line (Mac Terminal)**:
> `conda create -n r-dev -c r r-essentials`
> `conda activate r-dev`

To install packages from github, install R package `devtools`:
> `conda install r-devtools`

To ensure that dependency R package `units` [installs properly](https://github.com/r-quantities/units/issues/223), particularly for a conda environment:
> `conda install -c conda-forge udunits2`

Load R, and run the following:
> `devtools::install_github("ecological-cities/home2park@main")`

---

`sessionInfo()`

> R version 3.6.1 (2019-07-05)
> Platform: x86_64-apple-darwin13.4.0 (64-bit)
> Running under: macOS  10.16
>
> Matrix products: default
> BLAS/LAPACK: /Users/xpsong/opt/anaconda3/envs/r-dev/lib/R/lib/libRblas.dylib
>
> locale:
> [1] C/UTF-8/C/C/C/C
>
> attached base packages:
> [1] stats     graphics  grDevices utils     datasets  methods   base
>
> loaded via a namespace (and not attached):
> [1] compiler_3.6.1
