### 2021-09-04 #3609 @nickbearman
Thanks @xp-song for the notes re installation. I'm happy with this. @martinfleis any comments?
