### 2021-09-04 #3609 @nickbearman
Thanks @xp-song also for the tutorial link. This really helps.

On the tutorial itself, could you include the code you use to create the maps? I think this would be really useful for the tutorial.

Also, the example data is a bit slow to run on my machine. Would it be possible to trim it down a bit, so the code runs faster?
