### 2021-09-04 #3609 @nickbearman
Re the article, I have a conceptual question. You talk about
" the fact that supply from parks further away are reduced"
Which I think I understand (i.e. the closer people are to a park, the more likely they are to use it).
Could you rephrase it so we talk about people using the park, rather than the park being supplied to the people? I think this would make the article much easier to follow.

You also mention " By providing fine-grained measures for a variety of park attributes"
Can you list these attributes in the article? earlier on (2. Summarise at each park...) you mention attributes. Are these the same, or different?
