### 2021-09-05 e50da8c
subset example building data (city of Singapore) to 5% of full dataset, as the dataset is very large (and takes very long to process)
