### 2021-09-05 703c76f
add code in tutorial to visualise results using library(tmap)
