+++ vignettes/home2park.Rmd
@@ -253,12 +253,44 @@ buildings_pop_sgp$area_supply <- recre_supply(park_attribute = parks_sgp$area,
 
 # supply to all residents per building
 buildings_pop_sgp$area_supplytopop <- buildings_pop_sgp$area_supply * buildings_pop_sgp$popcount
-
 ```
 
 ```{r include = FALSE}
 rm(m_dist)
 ```
+
+We can visualise the supply of park area to building residents using `library(tmap)`:  
+
+```{r code to plot using building-level metrics using tmap, eval = FALSE}
+
+# convert buildings to points (centroids) for plotting
+buildings_pop_sgp <- buildings_pop_sgp %>% 
+  sf::st_centroid() %>%
+  dplyr::mutate(across(everything(), as.vector)) %>% # remove attributes from columns for plotting
+  dplyr::mutate(across(.cols = contains("area"), # convert from m2 to km2
+              .fns = function(x) x*1e-6))
+
+
+tmap::tmap_mode("view")
+
+tmap::tm_basemap("Esri.WorldGrayCanvas") +
+  tmap::tm_shape(parks_sgp) +
+    tmap::tm_polygons(col = "#33a02c",
+                      alpha = 0.6,
+                      border.col = "grey50", 
+                      border.alpha = 0.5) +
+  tmap::tm_shape(buildings_pop_sgp) +
+    tmap::tm_dots(title = "Supply of park area (km<sup>2</sup>)",
+                  col = "area_supplytopop", 
+                  border.col = "transparent",
+                  palette = viridis::viridis(5),
+                  style = "quantile",
+                  size = 0.01,
+                  alpha = 0.7,
+                  showNA = FALSE)
+
+```
+
 
 ```{r, echo=FALSE, fig.align = "center", out.width="100%", fig.cap = paste0("Screenshot: Supply of park area to building residents in Singapore based on OSM data (2020). Each building is denoted as a point (a random subset is shown). The color palette is binned according to the quantile values.")}
 knitr::include_graphics("../man/figures/README-supply-parkarea-to-building-residents.png") 
@@ -305,12 +337,12 @@ buildings_subzones <- buildings_subzones %>%
 # join information to pop_2020, calculate supply per capita
 pop_2020 <- pop_2020 %>%
   dplyr::left_join(buildings_subzones) %>%
-  dplyr::mutate(area_supplyperpop = area_supplytopop_sum / pop_count) %>%
+  dplyr::mutate(area_supplyperpop = area_supplytopop_sum / pop_count * 1e-6) %>% # convert to km2
   dplyr::mutate(area_supplyperpop = ifelse(is.infinite(area_supplyperpop), NA, area_supplyperpop)) # make infinite NA
 
 ```
 
-Calculate the per capita park area for each census unit:
+Calculate the (conventional) per capita park area for each census unit:
 
 ```{r include = FALSE}
 sf::st_agr(pop_2020) = "constant"
@@ -336,7 +368,7 @@ subzones_parks <- subzones_parks %>%
 pop_2020 <- pop_2020 %>%
   dplyr::left_join(subzones_parks) %>%
   dplyr::mutate(parkarea_m2 = ifelse(is.na(parkarea_m2), 0, parkarea_m2)) %>% # make NA 0
-  dplyr::mutate(parkperpop_m2 = parkarea_m2 / pop_count) %>%
+  dplyr::mutate(parkperpop_m2 = parkarea_m2 / pop_count * 1e-6) %>% # convert to km2
   dplyr::mutate(parkperpop_m2 = ifelse(is.infinite(parkperpop_m2), # make infinite NA
                                 NA, parkperpop_m2)) %>%
   dplyr::mutate(parkperpop_m2 = ifelse(is.nan(parkperpop_m2), # make NaN NA
[…]
