### 2021-09-05 #3609 @xp-song
> Thanks @xp-song also for the tutorial link. This really helps.
>
> On the tutorial itself, could you include the code you use to create the maps? I think this would be really useful for the tutorial.
>

Added in commit [703c76f](https://github.com/ecological-cities/home2park/commit/703c76f0858bec4a2b1bf5db5f35aa1903c89b57)

> Also, the example data is a bit slow to run on my machine. Would it be possible to trim it down a bit, so the code runs faster?

Amended example building dataset in commit [e50da8c](https://github.com/ecological-cities/home2park/commit/e50da8c9c35acdf9f9ec9bc9569db8e23ced809c)
