+++ JOSS-article/paper.md
@@ -45,7 +45,7 @@ See the [package website](https://ecological-cities.github.io/home2park/) for de
 
 # Statement of Need
 
-Numerous metrics have been used to measure the spatial provision of parks in cities. These include summaries of the park provision ratio (e.g. per capita park area) or ‘service radius’ (a distance buffer around parks) within geographical areal units such as census blocks or administrative boundaries [@Tan2017]. Such metrics have been used to investigate park use and accessibility, as well as social-cultural issues related to spatial equity and environmental justice [@Sister2010; @Tan2017]. However, the presence of artificial boundaries gives rise to the ‘modifiable areal unit problem’, where differences in spatial scale can result in considerable variation in the value of summarised metrics [@Sister2010]. Since urban development typically occurs at fine spatial scales, metrics at the scale of individual buildings can help overcome such limitations [@Gao2017]. Finally, most metrics of park provision do not differentiate between the wide variety of parks that would likely affect usage by different groups of people. For example, other than park area, other features such as waterfronts, open lawns, natural vegetation and trails each relate to different types of park use and user groups [@Song2020a; @Song2020b]. A more holistic measure of park provision should thus include a variety of park attributes that are important for recreation and leisure. By providing fine-grained measures for a variety of park attributes, we hope that ``home2park`` will contribute toward a more nuanced understanding of recreation that will improve the planning and design of parks and homes across the city.
+Numerous metrics have been used to measure the spatial provision of parks in cities. These include summaries of the per capita park area (i.e. park provision ratio) or ‘service radius’ (a distance buffer) around parks within geographical areal units such as census blocks or administrative boundaries [@Tan2017]. Such metrics have been used to investigate park use and accessibility, as well as social-cultural issues related to spatial equity and environmental justice [@Sister2010; @Tan2017]. However, the presence of artificial boundaries gives rise to the ‘modifiable areal unit problem’, where differences in spatial scale can result in considerable variation in the value of summarised metrics [@Sister2010]. Since urban development typically occurs at fine spatial scales, metrics at the scale of individual buildings provided by ``home2park`` can help overcome such limitations [@Gao2017]. Finally, most metrics of park provision do not differentiate between the wide variety of parks that would likely affect usage by different groups of people. Other than park area, features such as waterfronts, open lawns, natural vegetation, trails and fitness amenities each relate to different types of park use and user groups [@Song2020a; @Song2020b]. A more holistic measure of park provision should thus include a variety of park attributes that are important for recreation and leisure. ``home2park`` enables the user to summarise a variety of spatial attributes such as points (e.g. playgrounds, sports and fitness amenities), lines (e.g. cycling and walking trails) and rasters (e.g. land cover types) at each park. Customisable parameters provide the user with flexibility when summarising specific attributes, such as specifying a minimum patch size for a land cover class, or including spatial features within a certain distance beyond park boundaries. By providing fine-grained measures for a variety of park attributes, we hope that ``home2park`` will contribute toward a more nuanced understanding of recreation that will improve the planning and design of parks and homes across the city.
 
 ## State of the Field
 
