### 2021-09-06 #3609 @martinfleis
> Thanks @xp-song for the notes re installation. I'm happy with this. @martinfleis any comments?

If you're happy I am fine with that.
