### 2021-09-06 #3609 @aelissa
Just started looking at this.
Re the installation in the Vignette you suggest to run `devtools::install_github("ecological-cities/home2park")`. However,  while devtools::install_github uses master as default branch reference, your default is main. I would advise @xp-song to amend the installation command adding the correct ref `devtools::install_github("ecological-cities/home2park", ref="main")`
