### 2021-09-07 8930b7c
add 'ref = main' argument in devtools::install_github() function
