+++ README.Rmd
@@ -32,7 +32,7 @@ Assess the spatial provision of urban parks to residential buildings city-wide. 
 Install the development version of `home2park` from GitHub:
 
 ```{r, eval = FALSE}
-devtools::install_github("ecological-cities/home2park")
+devtools::install_github("ecological-cities/home2park", ref = "main")
 ```
 
 ## Setup
