+++ index.Rmd
@@ -31,7 +31,7 @@ knitr::opts_chunk$set(
 Install the development version of `home2park` from GitHub:
 
 ```{r, eval = FALSE}
-devtools::install_github("ecological-cities/home2park")
+devtools::install_github("ecological-cities/home2park", ref = "main")
 ```
 
 ## Setup
