+++ README.md
@@ -26,7 +26,7 @@ demonstrations of how the package may be used.
 Install the development version of `home2park` from GitHub:
 
 ``` r
-devtools::install_github("ecological-cities/home2park")
+devtools::install_github("ecological-cities/home2park", ref = "main")
 ```
 
 ## Setup
@@ -129,15 +129,15 @@ head(buildings_pop_sgp)
 #> Simple feature collection with 6 features and 1 field
 #> Geometry type: POLYGON
 #> Dimension:     XY
-#> Bounding box:  xmin: 103.8173 ymin: 1.460475 xmax: 103.8399 ymax: 1.46182
+#> Bounding box:  xmin: 103.8412 ymin: 1.297955 xmax: 103.8968 ymax: 1.426561
 #> Geodetic CRS:  WGS 84
-#>     popcount                       geometry
-#> 1 218.885435 POLYGON ((103.8179 1.46182,...
-#> 2  39.797352 POLYGON ((103.8173 1.461277...
-#> 3   9.219688 POLYGON ((103.8375 1.461288...
-#> 4   9.949338 POLYGON ((103.8176 1.461096...
-#> 5  85.282118 POLYGON ((103.839 1.461289,...
-#> 6 218.885435 POLYGON ((103.8176 1.461549...
+#>    popcount                       geometry
+#> 1 32.367111 POLYGON ((103.8412 1.420676...
+#> 2 20.880897 POLYGON ((103.8808 1.320019...
+#> 3  2.311594 POLYGON ((103.8643 1.320283...
+#> 4 56.747406 POLYGON ((103.8504 1.303723...
+#> 5  6.368404 POLYGON ((103.8516 1.426561...
+#> 6  9.502521 POLYGON ((103.8966 1.298226...
 ```
 
 <br>
