### 2021-09-07 #3609 @xp-song
> Just started looking at this.
> Re the installation in the Vignette you suggest to run `devtools::install_github("ecological-cities/home2park")`. However, while devtools::install_github uses master as default branch reference, your default is main. I would advise @xp-song to amend the installation command adding the correct ref `devtools::install_github("ecological-cities/home2park", ref="main")`

@aelissa Thanks for pointing this out, I've amended this is commit [8930b7c](https://github.com/ecological-cities/home2park/commit/8930b7cd080d396614af63e758b1eb86943d1712)
