### 2021-09-13 388cad1
rectify error that didn't allow specified column (custom name) to be used in rasterise_pop()
- pointed out in https://github.com/openjournals/joss-reviews/issues/3609#issuecomment-917699930
