### 2021-09-13 b229bb3
add documentation specifying the date format for get_building_osm(), and other functions that download OSM data
- pointed out in https://github.com/openjournals/joss-reviews/issues/3609#issuecomment-917699930
