### 2021-09-13 21c3c3b
amend documentation to clarify that inputs to pop_dasymap() are for a single year and requires extraction of list element(s) from intermediate datasets.
- pointed out in https://github.com/openjournals/joss-reviews/issues/3609#issuecomment-917699930
