+++ R/rasterise_pop.R
@@ -4,8 +4,8 @@
 #'into population density grid (raster) for subsequent analyses.
 #'Each polygon's count is divided by its area, such that the integrated
 #'density over each census block should equal the original count.
-#'Census data for multiple years can be processed (one raster per year).
-#'The column name for the `year` must be specified, even if data does not contain multiple years
+#'Census data for multiple years can be processed to output a list of rasters (one raster per year).
+#'The column name for the `'year'` in `sf` must be specified, even if data does not contain multiple years
 #'(combined data across multiple years must be in the 'long' format).
 #'
 #'@param sf `sf`polygons containing the population census data.
@@ -19,7 +19,8 @@
 #'
 #'@return List containing (1) rasterised population count per census block,
 #'(2) rasterised population density per census block, and (3) processed polygons used to generate
-#'these rasters. Named `pop_count_rasters`, `pop_density_rasters` and `pop_polygons`, respectively. List is nested if multiple years are present (one sub-list for each census year).
+#'these rasters. Named `pop_count_rasters`, `pop_density_rasters` and `pop_polygons`, respectively.
+#'List is nested if multiple years are present (one sub-list for each census year).
 #'Zero values for population counts/density are converted to `NA`. Intermediate (raster) files are exported to `tempdir()` for further processing.
 #'
 #'@import sf
