+++ man/pop_dasymap.Rd
@@ -38,8 +38,10 @@ Raster of population density, based on relative density values of the land defin
 }
 \description{
 Distribute population data per census block across inhabitable land as defined by \code{land_relative_density}.
-Data provided should be a snapshot of the population for a single time period (e.g. a year).
-All input data should have a similar projected coordinate reference system
+Raster data provided should be a snapshot of the population for a single time period (e.g. a year),
+created with other package functions (e.g. \code{rasterise_pop()}, \code{rasterise_buildings()}).
+If there are multiple years present in the intermediate datasets, the user will have to extract the relevant list element for
+the single year of interest (see examples). All input data should have a similar projected coordinate reference system
 specific to the target area.
 }
 \examples{
