### 2021-09-13 #3609 @xp-song
> I assessed the functionalities both by following the tutorial with the example data and by using some data for England. The tutorial is reproducible and the functions easily applicable to other data. Overall, I think it is a very useful package for analyzing parks' provision. @xp-song just a few comments:
>
> Re Functionality
>
> on rasterise_pop it seems that the population column needs to be named "pop count" to work. Try:
> `pop_sgp<-dplyr::rename(pop_sgp,pop=pop_count)`
>
> `pop_rasters <- rasterise_pop(pop_sgp, census_block = "subzone_n", pop_count = "pop")`
>
> Re Documentation
>
> In the documentation of get_building_osm you might want to specify the date's format "%y%m%d"
>
> There is a typo in rasterise_pop documentation at Description "such that the the integrated density"
>
> In pop_dasymap and the rasterising functions I think you should clarify a bit more how the output data structure includes time. My understanding it that you have one layer for each year. This needs to be made more explicit in the rasterising functions documentations as well as in pop_dasymap -- if you do not rely on external raster but on those created with the pkg functions -- it should be made clearer that you need to pick the element of a list of rasters related to the year of interest.
>
> I will look at the software paper and finalize my review in the next couple of days tops @martinfleis

Thank you! I've made the respective amendments in the linked commits above
