### 2021-09-17 #3609 @nickbearman
> Thanks @nickbearman for raising some good points. Some notes and responses:
>
> > Re the article, I have a conceptual question. You talk about
> > " the fact that supply from parks further away are reduced"
> > Which I think I understand (i.e. the closer people are to a park, the more likely they are to use it).
> > Could you rephrase it so we talk about people using the park, rather than the park being supplied to the people? I think this would make the article much easier to follow.
>
> Research on park provision tends to differentiate between '_supply_' (provision which may or may not be used), '_demand_', as well as actual '_use_'. The fact that parks are near to people's homes may not necessarily mean that people are actively '_using_' these places (which may be affected by factors other than distance). For e.g., desirability, cleanliness (on the supply side), or people's demographics/preferences such as whether they have children or pets (on the demand side) all affect actual park '_use_'. This is why the article/package documentation focuses on park '_supply_'.
>
> > You also mention " By providing fine-grained measures for a variety of park attributes"
> > Can you list these attributes in the article? earlier on (2. Summarise at each park...) you mention attributes. Are these the same, or different?
>
> I've amended the article text to try to clarify what I mean by "park attributes" in commit [55b1e77](https://github.com/ecological-cities/home2park/commit/55b1e777a7750800ccb9945ffa72c49e14976db9), and have also provided some examples. Thanks for helping to make the article clearer!

Thanks very much @xp-song this really does make things clearer
