### 2021-09-20 6ab39cd
update version no for JOSS publication
