+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: home2park
 Title: Spatial Provision of Urban Parks
-Version: 0.1.0
+Version: 0.1.1
 Authors@R: 
     c(person(given = "Xiao Ping",
              family = "Song",
