### 2021-09-20 #3609 @xp-song
@nickbearman @aelissa @martinfleis Thank you very much for your thoughtful review and comments. Here are requested information:

> * [x]  Make a new tagged release of the software, and list the version tag of the archived version here.

https://github.com/ecological-cities/home2park/releases/tag/v0.1.1

> * [x]  Archive the reviewed software in Zenodo or a similar service

https://zenodo.org/record/5516698#.YUfmNi0RppQ

> * [x ]  Please list the DOI of the archived version here.

10.5281/zenodo.5516698
