### 2021-09-20 #3609 @kthyng
I see your zenodo archive is made and in place, your version is up to date, and your paper looks great! We can proceed!
